pyproject.toml
[project]
name = "dsci531-2026-hw4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "python-docx>=1.2.0",
    "requests>=2.33.1",
]
